make all endpoints plural
