+++ facility_info/routers/proposals.py
@@ -7,3 +7,7 @@ from .proposal_data import proposals
 @router.get('/')
 def get_all_proposals():
     return {'proposals': proposals}
+
+@router.get('/{proposal_number}')
+def get_proposal_number(proposal_number: int):
+    return {'proposal': proposals[str(proposal_number)]}
