+++ facility_info/routers/users.py
@@ -5,6 +5,13 @@ router = APIRouter()
 user_info_dict = {'jdoe':{'username': 'jdoe',
     'globus_email': 'jdoe@bnl.gov'}}
 
+@router.get("/{username_or_email}")
+def get_user_info(username_or_email):
+    for user, info in user_info_dict.items():
+        if username_or_email in info.values():
+            return info
+    return None
+
 @router.get("/")
 def get_all_usernames():
     all_names = []
